Made function that creates 16 x 16 grid of square divs
To create 16 x 16 I used css grid, it only requires 2 parameters in order to create a grid row and column.

Function that creates grid, function grid(row,column). It creates rows and columns and then fills it up with square divs.

*
The way css properties are synchronized with javascript is by root: in css, it is less of a mess than to use DOM manipulation in javascript.